The most significant changes involve the addition of a new `RowVersion` property to the `Department` model in both the `SchoolContextModelSnapshot.cs` and `Department.cs` files. This property is used as a concurrency token and is updated every time a row is updated in the database. It is also decorated with the `[Timestamp]` attribute, indicating that it will be included in the `Where` clause of `UPDATE` and `DELETE` commands, to check for optimistic concurrency.
A new `DepartmentsController.cs` file has been added, which is responsible for handling HTTP requests related to departments. This includes methods for displaying a list of departments, displaying details of a specific department, creating a new department, editing an existing department, and deleting a department.

A new migration has been added in the `20240317234411_RowVersion.cs` file, which adds a new `RowVersion` column to the `Department` table in the database. The model snapshot in the `20240317234411_RowVersion.Designer.cs` file has been updated to reflect these changes.

Finally, new view files have been added (`Create.cshtml`, `Delete.cshtml`, `Details.cshtml`, `Edit.cshtml`, and `Index.cshtml`) which define the HTML that is sent to the client's browser for various department-related requests.

List of changes:

1. Addition of `RowVersion` property to `Department` model in `SchoolContextModelSnapshot.cs` and `Department.cs` files.
2. Addition of `DepartmentsController.cs` file with methods for handling department-related HTTP requests.
3. Addition of new migration in `20240317234411_RowVersion.cs` file to add `RowVersion` column to `Department` table in the database.
4. Update of model snapshot in `20240317234411_RowVersion.Designer.cs` file to reflect database changes.
5. Addition of new view files (`Create.cshtml`, `Delete.cshtml`, `Details.cshtml`, `Edit.cshtml`, and `Index.cshtml`) for department-related requests.